LIBFCREPO-1265. Moved validation functions to plastron.validation.rules.
https://umd-dit.atlassian.net/browse/LIBFCREPO-1265
