+++ plastron-models/src/plastron/models/poster.py
@@ -4,7 +4,7 @@ from plastron.models.umd import PCDMObject
 from plastron.namespaces import dcterms, dc, edm, bibo, geo, umd, umdtype
 from plastron.rdfmapping.decorators import rdf_type
 from plastron.rdfmapping.descriptors import ObjectProperty, DataProperty
-from plastron.validation import is_edtf_formatted, is_handle
+from plastron.validation.rules import is_edtf_formatted, is_handle
 
 
 @rdf_type(bibo.Image, umd.Poster)
