+++ plastron-models/tests/test_validation.py
@@ -1,14 +1,18 @@
+from email.message import Message
 from unittest.mock import Mock, patch
 from urllib.error import HTTPError
 
 import httpretty
 import pytest
 from httpretty import GET
-from rdflib import URIRef
+from rdflib import URIRef, Literal
 
+from plastron.namespaces import rdfs
 from plastron.rdf.rdf import RDFObjectProperty
-from plastron.validation import ValidationError, is_edtf_formatted, is_handle
-from plastron.validation.rules import from_vocabulary, required
+from plastron.rdfmapping.descriptors import DataProperty
+from plastron.rdfmapping.resources import RDFResourceBase
+from plastron.validation import ValidationError
+from plastron.validation.rules import is_edtf_formatted, is_handle, is_from_vocabulary
 from plastron.validation.vocabularies import get_vocabulary
 
 
@@ -27,17 +31,16 @@ from plastron.validation.vocabularies import get_vocabulary
         # non-string values pass
         ([0], True),
         ([1.0], True),
-        # only need one non-empty string to pass
-        (['foo', ''], True)
+        # any empty string fails
+        (['foo', ''], False),
     ]
 )
 def test_required(values, expected):
-    assert required(values) is expected
+    class SimpleResource(RDFResourceBase):
+        label = DataProperty(rdfs.label, required=True)
 
-
-def test_not_required():
-    # no values but not actually required passes
-    assert required([], False) is True
+    obj = SimpleResource(label=[Literal(str(v)) for v in values])
+    assert obj.is_valid == expected
 
 
 @pytest.mark.parametrize(
@@ -94,7 +97,8 @@ def test_not_handle(handle):
 def test_from_vocabulary(value, vocab_uri, expected):
     prop = RDFObjectProperty()
     prop.values = [URIRef(value)]
-    assert from_vocabulary(prop, vocab_uri) is expected
+    fn = is_from_vocabulary(vocab_uri)
+    assert fn(URIRef(value)) == expected
 
 
 @patch('plastron.validation.vocabularies.Graph')
@@ -112,7 +116,7 @@ def test_vocabulary_file_not_found(MockGraph):
 @patch('plastron.validation.vocabularies.Graph')
 def test_remote_vocab_error(MockGraph):
     mock_graph = Mock()
-    mock_graph.parse.side_effect = HTTPError('http://example.org/foo/', 503, '', {}, None)
+    mock_graph.parse.side_effect = HTTPError('http://example.org/foo/', 503, '', Message(), None)
     MockGraph.return_value = mock_graph
     # failure to retrieve the vocabulary over HTTP should
     # raise a ValidationError
@@ -146,6 +150,5 @@ def test_remote_vocab_308_redirect(shared_datadir):
         body=(shared_datadir / 'form.json').read_text(),
         content_type='application/ld+json'
     )
-    prop = RDFObjectProperty()
-    prop.values = [URIRef('http://vocab.lib.umd.edu/form#slides_photographs')]
-    assert from_vocabulary(prop, 'http://vocab.lib.umd.edu/form')
+    fn = is_from_vocabulary('http://vocab.lib.umd.edu/form')
+    assert fn(URIRef('http://vocab.lib.umd.edu/form#slides_photographs'))
